Group-Project-AI-Engine/SPRINT THREE/financial_analyst_crew/requirements.txt
pandas
numpy
yfinance
scikit-learn
tensorflow
matplotlib
flask
flask-cors
langchain-groq
langchain-ollama
langchain-openai
crewai
pydantic
firebase-admin
python-dotenv
PyYAML
gunicorn
